Revamped the service pages. Removed the home alone hero and all day care - was never used by cients.
I updated my rates for a desired, consistend 5 day, or twice a day schedule, with discounts to attract clients with 2-4 pets.

I include a page that showcased my veterinary experience as well as a list of pets I saw - this is to improve SEO so my website would hopefully gain clients with these pets.

Need to add content to FAQs page, fix hyperlink to the vet experiene page, and add photos to make the site more attractive.